requirements.txt
Django==4.0.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.22.1
drf-spectacular-sidecar==2022.6.1
pika==1.3.1
Pillow==9.1.1
PyJWT==2.4.0
python-decouple==3.6
PyYAML==6.0
requests==2.28.0
psycopg2-binary==2.9.5
gunicorn==20.1.0
django-extensions==3.2.1
